* move DocumentSnapshot to model.ts * move FileSnapshotContent to model.ts
