+++ server/aws-lsp-codewhisperer/src/shared/models/model.ts
@@ -37,3 +37,29 @@ export enum SuggestionType {
     EDIT = 'EDIT',
     COMPLETION = 'COMPLETION',
 }
+
+/**
+ * Represents a snapshot of a document at a specific point in time
+ */
+export interface DocumentSnapshot {
+    /** URI of the document */
+    readonly filePath: string
+    /** Size of the snapshot content in bytes */
+    readonly size: number
+    /** Timestamp when the snapshot was taken */
+    readonly timestamp: number
+    /** Content of the document at the time of snapshot */
+    readonly content: string
+}
+
+/**
+ * Represents a snapshot content of a file at a specific point in time
+ */
+export interface FileSnapshotContent {
+    /** URI of the file */
+    readonly filePath: string
+    /** Content of the file */
+    readonly content: string
+    /** Timestamp when the snapshot was taken */
+    readonly timestamp: number
+}
